+++ test/test_helper.rb
@@ -85,7 +85,7 @@ class Minitest::Test
   def call_method(method, field, options)
     if enumerable?
       Hash[@users.group_by_period(method, **options) { |u| u.send(field) }.map { |k, v| [k, v.size] }]
-    elsif sqlite? && (method == :quarter || options[:time_zone] || options[:day_start] || options[:week_start] || Groupdate.week_start != :sun || (Time.zone && options[:time_zone] != false))
+    elsif sqlite? && (method == :quarter || options[:time_zone] || options[:day_start] || (Time.zone && options[:time_zone] != false))
       error = assert_raises(Groupdate::Error) { User.group_by_period(method, field, **options).count }
       assert_includes error.message, "not supported for SQLite"
       skip